requirements.txt
asgiref==3.9.1
Django==5.2.4
django-cors-headers==4.7.0
django-restframework==0.0.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
gunicorn==23.0.0
packaging==25.0
pillow==11.3.0
PyJWT==2.10.1
sqlparse==0.5.3
tzdata==2025.2
python-dotenv==1.0.1
